Use explicit type in switch expression in ConsumerTest (#87)
Change the switch expression inConsumerTest#validateActiveMqMessage to
declare the type explicitly instead of using var. When using var, and
compiling inside Gitpod via the RedHat Language Support for Java
extension, we get an error trying to run the test. Here is a snippet of
the stack trace:

Caused by: java.lang.VerifyError: Bad type on operand stack Exception Details:
  Location:
    org/kiwiproject/dropwizard/activemq/internal/ConsumerTest.validateActiveMqMessage(Lorg/kiwiproject/dropwizard/activemq/ActiveMqMessage;Ljava/lang/String;Ljava/lang/String;)V @220: invokestatic
  Reason:
    Type 'java/lang/Object' (current frame, stack[0]) is not assignable to 'java/lang/String'

This appears to be a problem in the Eclipse compiler, which the RedHat
Java extension is using. This does not happen when compiling/running
tests using Maven or IntelliJ.

So, for now I have changed the switch expression to declare the local
variable as String instead of using var. I left the TODO in place and
added some additional information and links.

Also, removed the block of commented out code. Not sure why I had left
that in there...